# ZTS: Optimize KSM on Linux and remove it for FreeBSD