**This patchset does:**

Optimize KSM settings for Linux and disable it for FreeBSD

I reemoved the "Fix FreeBSD problem" with [#16822](https://github.com/mcmilk/zfs/actions/runs/14619248183) ... this should maybe go into a seperat PR.

| System | [Before](https://github.com/openzfs/zfs/actions/runs/14631209833/usage) | [New](https://github.com/openzfs/zfs/actions/runs/14633648401/usage) | [latest](https://github.com/mcmilk/zfs/actions/runs/14658449095/usage) |
| --- | --- | --- | --- |
|almalinux8|3h 25m 47s|3h 17m 23s|3h 18m 9s
|almalinux9|3h 1m 9s|2h 48m 18s|2h 46m 41s
|debian11|3h 31m 43s|3h 12m 58s|3h 16m 35s
|debian12|2h 58m 27s|2h 33m 33s|2h 38m 44s
|fedora40|3h 2m 11s|2h 43m 19s|2h 48m 50s
|fedora41|3h 0m 8s|2h 41m 21s|2h 37m 37s
|fedora42|2h 55m 35s|2h 41m 40s|2h 41m 44s
|freebsd13-4r|2h 15m 55s|1h 43m 33s|1h 43m 22s
|freebsd14-2r|2h 16m 32s|1h 44m 40s|1h 46m 48s
|freebsd15-0c|3h 33m 55s|2h 20m 53s|2h 16m 30s
|ubuntu20|3h 9m 18s|2h 58m 12s|2h 58m 33s
|ubuntu22|3h 44m 4s|3h 28m 41s|3h 37m 23s
|ubuntu24|3h 6m 6s|2h 50m 37s|2h 47m 55s
|total|1d 16h 1m 13s|1d 11h 5m 34s|1d 11h 19m 18s

### Motivation and Context

Some ZTS test runs fail because of memory pressure. This PR tries to fix this.

### Description

All commits have the needed infos ;-)

### How Has This Been Tested?

Tried here and there, until the result was stable and the times went down ;)

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
<!--- If your change is a performance enhancement, please provide benchmarks here. -->
<!--- Please think about using the draft PR feature if appropriate -->

### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Performance enhancement (non-breaking change which improves efficiency)
- [ ] Code cleanup (non-breaking change which makes code smaller or more readable)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Library ABI change (libzfs, libzfs\_core, libnvpair, libuutil and libzfsbootenv)
- [ ] Documentation (a change to man pages or other documentation)

### Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the OpenZFS [code style requirements](https://github.com/openzfs/zfs/blob/master/.github/CONTRIBUTING.md#coding-conventions).
- [ ] I have updated the documentation accordingly.
- [x] I have read the [**contributing** document](https://github.com/openzfs/zfs/blob/master/.github/CONTRIBUTING.md).
- [ ] I have added [tests](https://github.com/openzfs/zfs/tree/master/tests) to cover my changes.
- [x] I have run the ZFS Test Suite with this change applied.
- [x] All commit messages are properly formatted and contain [`Signed-off-by`](https://github.com/openzfs/zfs/blob/master/.github/CONTRIBUTING.md#signed-off-by).