@tonyhutter 2025-04-15
> Free kernel caches on Linux VMs via cron, to get more usable memory on them.

For Linux, you might be better off playing with some of the virtual memory tunables on either the runner or the VMs.  For example, have you tried setting `vm.min_free_kbytes` to higher values?