@amotin 2025-04-22
> > Could we commit the version updates independently before jumping into optimizations? I'd really like to see FreeBSD 15 to finally coming clean, while the VM management changes look less straightforward.
> 
> Should I create a seperate PR for this?

Yes, please.  I'd be happy to merge it once it reasonably pass CI, hopefully in not too long, and then we could return to this one.