@tonyhutter 2025-04-22
Just want to say that shaving 6 hours off the total run time is really impressive :+1: